Resolve issues
