+++ readme.md
@@ -1,4 +1,6 @@
-# [https://jorianp.github.io/project11](https://jorianp.github.io/project11)
-## Проектное задание 11
+# Проектное задание 11
+## [https://jorianp.github.io/project11](https://jorianp.github.io/project11)
 ### v. 1.0.1
 Спринт 11: Модули в JS, Webpack
+Webpack конфиг для production: npm run build
+                   development: npm run dev
