+++ src/index.js
@@ -1,4 +1,2 @@
-const serverUrl = NODE_ENV === 'development' ? 'http://praktikum.tk/cohort2' : 'https://praktikum.tk/cohort2';
-
 import '../scripts/script'
 import '../pages/index.css'
